requirements.txt
streamlit
pandas
python-docx
PyPDF2
docx2txt
